requirements.txt
streamlit
langchain-openai
python-dotenv
langgraph
amadeus
langchain_chroma
ipython
serpapi
google-search-results
reportlab
markdown2
bs4